Handle (some) quotes, multiple consecutive delimiters
Wherein I learn something about lifetimes and ownership: when readconf.c sees a quote, it modifies the underlying string to elide it. When I switched to borrows in 3085471, I complicated the ability of parse_line to make modifications to the string.

Wrote down some questions and tradeoffs I see for now, as well as an #[ignored] test that checks for the behavior demonstrated in some of the integration tests.