refactor: towards a more complete language
With the discovery that arguments can be variadic depending on the keyword,
we have to separate the keyword parsing from the part that consumes whatever
arguments are appropriate.

In order to preserve the "only allocate in the rare case that we have to modify"
poossibility (presuming some non-allocating case-insensitive comparison) we opt
for nested closures that each consume a variable amount of the token stream.