fix: propagate unmatched quote errors
However, we open ourselves up to a new class of failure this way by allowing
users of the `toks` stream to "return" via a side channel whatever state they
want at the front, including something that never appeared in the stream.

We've got some more work to do to preserve the "no leading delimiters" invariant.